pyproject.toml
[project]
name = "llm-experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=4.0.0",
    "python-dotenv>=1.2.1",
]
